Update ikala permissions to gradle structure
