+++ Ikala-Permissions/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'Ikala-Permissions'
+include ':Ikala-Core'
+project(':Ikala-Core').projectDir = new File('../../Ikala-Core')
