--- Ikala-Permissions/src/com/ikalagaming/permissions/DuplicateGroupException.java
@@ -1,17 +1,0 @@
-package com.ikalagaming.permissions;
-
-/**
- * Thrown when trying to create a permission group with a name that is already
- * used.
- *
- * @author Ches Burks
- *
- */
-public class DuplicateGroupException extends RuntimeException {
-
-	/**
-	 * Generated serial ID
-	 */
-	private static final long serialVersionUID = 5217138074629583815L;
-
-}
