--- Ikala-Permissions/src/com/ikalagaming/permissions/package-info.java
@@ -1,4 +1,0 @@
-/**
- * A group-based permissions system.
- */
-package com.ikalagaming.permissions;
