--- Ikala-Permissions/src/com/ikalagaming/permissions/resources/Permissions.properties
@@ -1,3 +1,0 @@
-INVALID_PERMISSIONS=Invalid permissions
-NON_EXISTANT_SUBPERMISSON=Sub-permission {} does not exist
-DUPLICATE_PERMISSION=Permission {} already exists
