+++ Ikala-Permissions/src/main/java/com/ikalagaming/permissions/DuplicateGroupException.java
@@ -1,0 +1,17 @@
+package com.ikalagaming.permissions;
+
+/**
+ * Thrown when trying to create a permission group with a name that is already
+ * used.
+ *
+ * @author Ches Burks
+ *
+ */
+public class DuplicateGroupException extends RuntimeException {
+
+	/**
+	 * Generated serial ID
+	 */
+	private static final long serialVersionUID = 5217138074629583815L;
+
+}
