+++ Ikala-Permissions/src/main/java/com/ikalagaming/permissions/EmptyGroupNameException.java
@@ -1,0 +1,17 @@
+package com.ikalagaming.permissions;
+
+/**
+ *
+ * Thrown when a permission group is created with a blank or null name.
+ *
+ * @author Ches Burks
+ *
+ */
+public class EmptyGroupNameException extends RuntimeException {
+
+	/**
+	 * Generated serial ID
+	 */
+	private static final long serialVersionUID = 7593339696547250731L;
+
+}
