+++ Ikala-Permissions/src/main/java/com/ikalagaming/permissions/PermissionPlugin.java
@@ -1,0 +1,13 @@
+package com.ikalagaming.permissions;
+
+import com.ikalagaming.plugins.Plugin;
+
+/**
+ * Lifecycle methods for permissions.
+ *
+ * @author Ches Burks
+ *
+ */
+public class PermissionPlugin extends Plugin {
+
+}
