+++ Ikala-Permissions/src/main/java/com/ikalagaming/permissions/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * A group-based permissions system.
+ */
+package com.ikalagaming.permissions;
