+++ Ikala-Permissions/src/main/resources/com/ikalagaming/permissions/Permissions.properties
@@ -1,0 +1,3 @@
+INVALID_PERMISSIONS=Invalid permissions
+NON_EXISTANT_SUBPERMISSON=Sub-permission {} does not exist
+DUPLICATE_PERMISSION=Permission {} already exists
